# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 PetStore API",
        "description": "L5 PetStore API description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/": {
            "get": {
                "description": "Home Page",
                "responses": {
                    "default": {
                        "description": "Welcome Page"
                    }
                }
            }
        },
        "/api/pets": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "description": "Displays all the pets",
                "operationId": "7b44ba1ad8a38f7865a2845fa9ebbfe3",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Pets in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pets"
                ],
                "description": "Create new pet",
                "operationId": "cfc3bb9eb41d8f40bcb678b5194c5a96",
                "parameters": [
                    {
                        "name": "species",
                        "in": "path",
                        "description": "Species of Pet being added",
                        "schema": {
                            "type": "varchar"
                        },
                        "example": "Dog"
                    },
                    {
                        "name": "breed",
                        "in": "path",
                        "description": "Breed of Pet species",
                        "schema": {
                            "type": "varchar"
                        },
                        "example": "Labrador"
                    },
                    {
                        "name": "description",
                        "in": "path",
                        "description": "Description of Pet",
                        "schema": {
                            "type": "text"
                        },
                        "example": "Brown with white spots"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Pet",
                        "schema": {
                            "type": "varchar"
                        },
                        "example": "Bruno"
                    },
                    {
                        "name": "age",
                        "in": "path",
                        "description": "Age of Pet",
                        "schema": {
                            "type": "int"
                        },
                        "example": 8
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation, Created a new pet"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/pets/id": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "description": "Displays Pet with specified id",
                "operationId": "f325eca2cb2e9d00e58892ee6528f833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pet",
                        "schema": {
                            "type": "int"
                        },
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns Pet corresponding to id"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pets"
                ],
                "description": "Create new pet",
                "operationId": "e53ab5a7da00b5b97e02ea68c65c9f9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pet you want to Update",
                        "schema": {
                            "type": "int"
                        },
                        "example": 7
                    },
                    {
                        "name": "species",
                        "in": "path",
                        "description": "Species of Pet",
                        "schema": {
                            "type": "varchar"
                        },
                        "example": "Dog"
                    },
                    {
                        "name": "breed",
                        "in": "path",
                        "description": "Breed of Pet species",
                        "schema": {
                            "type": "varchar"
                        },
                        "example": "Labrador"
                    },
                    {
                        "name": "description",
                        "in": "path",
                        "description": "Description of Pet",
                        "schema": {
                            "type": "text"
                        },
                        "example": "Brown with white spots"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Pet",
                        "schema": {
                            "type": "varchar"
                        },
                        "example": "Bruno"
                    },
                    {
                        "name": "age",
                        "in": "path",
                        "description": "Age of Pet",
                        "schema": {
                            "type": "int"
                        },
                        "example": 8
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pets"
                ],
                "description": "Delete a Pet",
                "operationId": "44f477ebce07350eebc44f7a3fddd8b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pet you want to Delete",
                        "schema": {
                            "type": "int"
                        },
                        "example": 4
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            }
        }
    }
}